+++ src/const.ts
@@ -1,3 +1,5 @@
+import { config } from "./config";
+
 /**
  * Whether the current platform is Windows.
  */
@@ -8,3 +10,8 @@ export const isWindows = /^win/.test(process.platform);
  * This is `rg.exe` on Windows and `rg` on other platforms.
  */
 export const binName = isWindows ? "rg.exe" : "rg";
+
+if (config.debug) {
+  console.log(`isWindows: ${isWindows}`);
+  console.log(`binName: ${binName}`);
+}